refactor(tui): colocate Screen::Config into tui/screens/config.rs
Pilot for issue #287 (Phase 2 — colocation): one Screen variant's behavior
is smeared across 5 per-screen registries (handle_key dispatch, HelpTopic::
for_screen, wheel_step, build_view_model's ScreenVm match, render_screen_vm's
paint dispatch, build_palette_items). Option B' keeps the registries'
exhaustive-match shape but moves each screen's handle_key_*/build_*_vm/
render_*_vm/*_palette_items bodies into one file, tui/screens/<screen>.rs,
one screen at a time.

This PR does exactly one screen — Screen::Config, the simplest/most
self-contained — to establish the pattern before migrating the other 10.

Moved into the new tui/screens/config.rs: handle_key_config and
adjust_config_field (as another impl AppState block, an already-established
pattern in this codebase), build_config_vm, render_config_vm,
config_palette_items, plus a HELP_TOPIC const for HelpTopic::for_screen's
arm. wheel_step has no dedicated Screen::Config arm (already falls through
to the shared default), so nothing moved there. Pure move — bodies and all
existing tests (across tests.rs, view_model.rs/palette.rs/render.rs's own
test modules) are unchanged; they exercise the moved logic through the
registries' dispatch, not by calling the relocated functions directly.

Deliberately left in place: open_config() and config_field_value() (both in
keys.rs) are shared beyond Config's own dispatch — open_config is called
from other screens' key handlers and the palette's cross-action;
config_field_value is also called from bg.rs. ConfigState/ConfigField/
ConfigVm type definitions stay in mod.rs/view_model.rs — the issue's scope
is function bodies, not type definitions.

render_close_button (render.rs) and key_item (palette.rs) widen from
private to pub(super) — the minimum needed for tui::screens::config (a
descendant of tui) to reach them, matching the existing pub(super) pattern
already used by render_top_bar for the same kind of cross-submodule call.

Part of #287
