+++ src/tui/palette.rs
@@ -274,7 +274,7 @@ fn palette_item(key: &str, label: &str, exec: PaletteExec, enabled: bool) -> Pal
     }
 }
 
-fn key_item(key: &str, label: &str, code: KeyCode, enabled: bool) -> PaletteItem {
+pub(super) fn key_item(key: &str, label: &str, code: KeyCode, enabled: bool) -> PaletteItem {
     palette_item(
         key,
         label,
@@ -330,7 +330,7 @@ fn build_palette_items(state: &AppState, screen: &Screen, mode: PaletteMode) -> 
         Screen::Diff(_) => diff_palette_items(state),
         Screen::Preview(_) => preview_palette_items(state),
         Screen::Help(_) => help_palette_items(),
-        Screen::Config(_) => config_palette_items(),
+        Screen::Config(_) => super::screens::config::config_palette_items(),
         Screen::Confirm(_) | Screen::Palette(_) => Vec::new(),
     };
     if mode == PaletteMode::Command {
@@ -611,14 +611,6 @@ fn preview_palette_items(_state: &AppState) -> Vec<PaletteItem> {
         key_item("y", "Copy gist URL", KeyCode::Char('y'), true),
         key_item("Y", "Copy file content", KeyCode::Char('Y'), true),
         key_item("q", "Back", KeyCode::Char('q'), true),
-    ]
-}
-
-fn config_palette_items() -> Vec<PaletteItem> {
-    vec![
-        key_item("Enter", "Toggle / increase value", KeyCode::Enter, true),
-        key_item("h/l", "Decrease / increase value", KeyCode::Char('l'), true),
-        key_item("Esc", "Close settings", KeyCode::Esc, true),
     ]
 }
